Update plugin to use the newer RPC methods

+++ src/lines.rs
@@ -28,8 +28,8 @@ impl Lines {
   /// Update Lines:
   ///   * `first` and `last` are zero-indexed (just as Lines itself)
   ///   * If `numreplaced` is zero, the lines were added before line `first`
-  pub fn update(&mut self, first: u64, last: u64, linedata: Vec<String>) {
-    let range = first as usize..last as usize;
+  pub fn update(&mut self, first: usize, last: usize, linedata: Vec<String>) {
+    let range = first..last;
     let _v = self.0.splice(range, linedata);
   }
 }
